requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
mysql-connector-python==9.0.0
mysqlconnect==1.0.0
numpy==2.1.1
pandas==2.2.3
PySide6==6.8.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
shiboken6==6.7.2
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.2.3
